arrumando
corrigindo alguns erros do "merge na mao", e adicionando algumas notas sobre o que devemos fazer na estrategia
